Refactor with CodeMetrics
