+++ core/src/main/java/co/com/sofka/core/issue/queries/QueryGetIssueById.java
@@ -1,4 +1,4 @@
-package co.com.sofka.core.issue.quieries;
+package co.com.sofka.core.issue.queries;
 
 import co.com.sofka.domain.generic.QueryGeneric;
 import com.fasterxml.jackson.annotation.JsonProperty;
